Add Scalingo utility functions and integrate into retrieve_companies DAG
